Improve consts move Sensor childclass
